requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.7.0
pydantic-settings==2.3.0
google-generativeai==0.7.0
slack-sdk==3.31.0
requests==2.32.0
python-dotenv==1.0.1
python-multipart==0.0.9
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.143.0
python-quickbooks==0.9.7
aiosqlite==0.20.0
